python/requirements.txt
schedule==1.2.0
python-dotenv==1.0.0
tse-utils